Use alternate generic function 'dump' to store results
